+++ Vault/src/utopia/vault/model/template/HasId.scala
@@ -2,6 +2,7 @@ package utopia.vault.model.template
 
 /**
  * Common trait for classes which specify a (database row) id
+ * @tparam Id Type of id of this instance
  * @author Mikko Hilpinen
  * @since 15/03/2024, v1.19
  */